make AngularScalarTuner general
